debug

+++ giraph-core/src/main/java/org/apache/giraph/utils/SigarUtil.java
@@ -1,10 +1,16 @@
 package org.apache.giraph.utils;
 
 import com.google.common.io.Resources;
+import org.apache.hadoop.conf.Configuration;
+import org.apache.hadoop.fs.FileSystem;
+import org.apache.hadoop.fs.Path;
+import org.apache.hadoop.io.IOUtils;
+import org.apache.hadoop.mapreduce.Mapper;
 import org.apache.log4j.Logger;
 import org.hyperic.sigar.Sigar;
 
 import java.io.*;
+import java.net.URI;
 import java.net.URL;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipFile;
@@ -16,7 +22,7 @@ import java.util.zip.ZipInputStream;
 public class SigarUtil {
 
 
-    public static Sigar getSigar() throws IOException{
+    public static Sigar getSigar(Mapper<?, ?, ?, ?>.Context context) throws IOException{
         final Logger LOG = Logger.getLogger(SigarUtil.class);
 
         String userHome = System.getProperty("user.home");
@@ -26,27 +32,51 @@ public class SigarUtil {
         if(!sigarFolder.exists()){
             sigarFolder.mkdir();
 
-            InputStream is = SigarUtil.class.getResourceAsStream("sigar_lib.zip");
-            System.out.println("SigarUtil.class.getCanonicalName()" + SigarUtil.class.getCanonicalName());
-            ZipInputStream zis = new ZipInputStream(is);
+            Configuration conf = context.getConfiguration();
+            FileSystem fileSystem = FileSystem.get(URI.create("hdfs://master:9000"), conf);
+            Path path = new Path("/libraries/libsigar-amd64-linux.so");
+            if(!fileSystem.exists(path)){
+                LOG.info(path.toString() + " does not exist!");
+                return null;
+            }
+            InputStream in = fileSystem.open(path);
+            File file = new File(sigarFolderName + "/libsigar-amd64-linux.so");
+            if(!file.exists()){
+                file.createNewFile();
+            }
+            OutputStream out = new FileOutputStream(file.getCanonicalFile());
+            IOUtils.copyBytes(in, out, 4096, true);
+
+            /*
+            //System.out.println(Resources.getResource("sigar_lib.zip"));
+            ///usr/local/lib/hadoop-1.2.1/tmp/mapred/local/taskTracker/hadoop/jobcache/job_201801121423_0006/jars/sigar_lib.zip
+            //File file = new File(Resources.getResource("sigar_lib.zip").getFile());
+            //InputStream is = SigarUtil.class.getResourceAsStream("sigar_lib.zip");
+            //System.out.println("SigarUtil.class.getCanonicalName()" + SigarUtil.class.getCanonicalName());
+            ZipFile zip = new ZipFile(file);
+            ZipInputStream zis = new ZipInputStream(new FileInputStream(file));
 
             ZipEntry entry;
             while ((entry = zis.getNextEntry()) != null) {
                 System.out.println("decompress file :" + entry.getName());
                 File outFile = new File(sigarFolderName + "/" + entry.getName());
+                BufferedInputStream bis = new BufferedInputStream(zip.getInputStream(entry));
                 BufferedOutputStream bos = new BufferedOutputStream(new FileOutputStream(outFile));
 
                 byte[] buffer = new byte[1024];
                 while (true) {
-                    int len = zis.read(buffer);
+                    int len = bis.read(buffer);
                     if (len == -1)
                         break;
                     bos.write(buffer, 0, len);
                 }
+                bis.close();
                 bos.close();
             }
             zis.close();
+            */
         }
+
 
         String seperator = null;
         if (OsCheck.getOperatingSystemType() == OsCheck.OSType.Windows)
